refactor(wip): Moves some code and tests around
