+++ src/on-chain-data/index.js
@@ -1,8 +1,11 @@
 // @flow
 
+import type { DataModelInterface, AdaptedTxResultInterface, AdaptedTxResultsInterface, KeystoreV3Interface } from '../interfaces/base-interfaces';
+import type { WTHotelIndexInterface } from '../interfaces/hotel-interfaces';
+import type { WTAirlineIndexInterface } from '../interfaces/airline-interfaces';
+
 import Utils from './utils';
 import Contracts from './contracts';
-import type { DataModelInterface, AdaptedTxResultInterface, AdaptedTxResultsInterface, KeystoreV3Interface } from '../interfaces/base-interfaces';
 import Wallet from '../wallet';
 import WTHotelIndex from './hotels/wt-index';
 import WTAirlineIndex from './airlines/wt-index';
@@ -36,10 +39,30 @@ class AbstractDataModel implements DataModelInterface {
   web3Utils: Utils;
   web3Contracts: Contracts;
 
+  _wtIndexCache: {[address: string]: WTHotelIndexInterface | WTAirlineIndexInterface};
+
+  /**
+   * Sets up Utils and Contracts with given web3 provider.
+   * Sets up gasCoefficient or gasMargin. If neither is provided,
+   * sets gasCoefficient to a default of 2.
+   */
+  constructor (options: DataModelOptionsType) {
+    this.options = options || {};
+    if (!this.options.gasMargin && !this.options.gasCoefficient) {
+      this.options.gasCoefficient = 2;
+    }
+    this.web3Utils = Utils.createInstance({
+      gasCoefficient: this.options.gasCoefficient,
+      gasMargin: this.options.gasMargin,
+    }, this.options.provider);
+    this.web3Contracts = Contracts.createInstance(this.options.provider);
+    this._wtIndexCache = {};
+  }
+
   /**
    * Returns an Ethereum backed Winding Tree index.
    */
-  getWindingTreeIndex (address: string): WTHotelIndex | WTAirlineIndex {
+  _indexFactory (address: string): WTHotelIndexInterface | WTAirlineIndexInterface {
     throw Error('Not implemented. Should be called on a subclass instance.');
   }
 
@@ -90,6 +113,16 @@ class AbstractDataModel implements DataModelInterface {
   }
 
   /**
+   * Returns an Ethereum backed Winding Tree index.
+   */
+  getWindingTreeIndex (address: string): WTHotelIndexInterface | WTAirlineIndexInterface {
+    if (!this._wtIndexCache[address]) {
+      this._wtIndexCache[address] = this._indexFactory(address);
+    }
+    return this._wtIndexCache[address];
+  }
+
+  /**
    * Returns a wallet instance for given JSON keystore.
    */
   createWallet (jsonWallet: KeystoreV3Interface): Wallet {
@@ -100,82 +133,28 @@ class AbstractDataModel implements DataModelInterface {
 }
 
 class HotelDataModel extends AbstractDataModel {
-  _wtIndexCache: {[address: string]: WTHotelIndex};
-
   /**
-   * Sets up Utils and Contracts with given web3 provider.
-   * Sets up gasCoefficient or gasMargin. If neither is provided,
-   * sets gasCoefficient to a default of 2.
-   */
-  constructor (options: DataModelOptionsType) {
-    super();
-    this.options = options || {};
-    if (!this.options.gasMargin && !this.options.gasCoefficient) {
-      this.options.gasCoefficient = 2;
-    }
-    this.web3Utils = Utils.createInstance({
-      gasCoefficient: this.options.gasCoefficient,
-      gasMargin: this.options.gasMargin,
-    }, this.options.provider);
-    this.web3Contracts = Contracts.createInstance(this.options.provider);
-    this._wtIndexCache = {};
-  }
-
-  /**
-   * Creates a configured AbstractDataModel instance.
+   * Creates a configured HotelDataModel instance.
    */
   static createInstance (options: DataModelOptionsType): AbstractDataModel {
     return new HotelDataModel(options);
   }
 
-  /**
-   * Returns an Ethereum backed Winding Tree index.
-   */
-  getWindingTreeIndex (address: string): WTHotelIndex {
-    if (!this._wtIndexCache[address]) {
-      this._wtIndexCache[address] = WTHotelIndex.createInstance(address, this.web3Utils, this.web3Contracts);
-    }
-    return this._wtIndexCache[address];
+  _indexFactory (address: string): WTHotelIndex {
+    return WTHotelIndex.createInstance(address, this.web3Utils, this.web3Contracts);
   }
 }
 
 class AirlineDataModel extends AbstractDataModel {
-  _wtIndexCache: {[address: string]: WTAirlineIndex};
-
   /**
-   * Sets up Utils and Contracts with given web3 provider.
-   * Sets up gasCoefficient or gasMargin. If neither is provided,
-   * sets gasCoefficient to a default of 2.
-   */
-  constructor (options: DataModelOptionsType) {
-    super();
-    this.options = options || {};
-    if (!this.options.gasMargin && !this.options.gasCoefficient) {
-      this.options.gasCoefficient = 2;
-    }
-    this.web3Utils = Utils.createInstance({
-      gasCoefficient: this.options.gasCoefficient,
-      gasMargin: this.options.gasMargin,
-    }, this.options.provider);
-    this.web3Contracts = Contracts.createInstance(this.options.provider);
-    this._wtIndexCache = {};
-  }
-
-  /**
-   * Creates a configured AbstractDataModel instance.
+   * Creates a configured AirlineDataModel instance.
    */
   static createInstance (options: DataModelOptionsType): AbstractDataModel {
     return new AirlineDataModel(options);
   }
 
-  /**
-   * Returns an Ethereum backed Winding Tree index.
-   */
-  getWindingTreeIndex (address: string): WTAirlineIndex {
-    if (!this._wtIndexCache[address]) {
-      this._wtIndexCache[address] = WTAirlineIndex.createInstance(address, this.web3Utils, this.web3Contracts);
-    }
-    return this._wtIndexCache[address];
+  _indexFactory (address: string): WTAirlineIndex {
+    return WTAirlineIndex.createInstance(address, this.web3Utils, this.web3Contracts);
   }
 }
