--- test/wt-libs/on-chain-data/airlines/airline-index.spec.js
@@ -1,60 +1,0 @@
-import { assert } from 'chai';
-import sinon from 'sinon';
-import { WtJsLibs } from '../../../../src/index';
-import { AirlineDataModel } from '../../../../src/on-chain-data';
-import OffChainDataClient from '../../../../src/off-chain-data-client';
-import { AIRLINE_SEGMENT_ID } from '../../../../src/on-chain-data/constants';
-
-describe('WTLibs.WTAirlineIndex', () => {
-  describe('createInstance', () => {
-    let createDataModelSpy;
-
-    beforeEach(() => {
-      createDataModelSpy = sinon.spy(AirlineDataModel, 'createInstance');
-    });
-
-    afterEach(() => {
-      createDataModelSpy.restore();
-    });
-
-    it('should initialize data model', () => {
-      const libs = WtJsLibs.createInstance({ segment: AIRLINE_SEGMENT_ID });
-      assert.isDefined(libs.dataModel);
-      assert.equal(createDataModelSpy.callCount, 1);
-    });
-
-    it('should pass data model options', () => {
-      const libs = WtJsLibs.createInstance({
-        segment: AIRLINE_SEGMENT_ID,
-        dataModelOptions: {
-          random: '1234',
-        },
-      });
-      assert.isDefined(libs.dataModel);
-      assert.equal(createDataModelSpy.firstCall.args[0].random, '1234');
-    });
-  });
-
-  describe('getOffChainDataClient', () => {
-    it('should return OffChainDataClient', () => {
-      const libs = WtJsLibs.createInstance({
-        segment: AIRLINE_SEGMENT_ID,
-        dataModelOptions: {
-          random: '1234',
-        },
-        offChainDataOptions: {
-          adapters: {
-            'in-memory': {
-              create: () => {
-                return true;
-              },
-            },
-          },
-        },
-      });
-      const adapter = libs.getOffChainDataClient('in-memory');
-      assert.isDefined(adapter);
-      OffChainDataClient._reset();
-    });
-  });
-});
