+++ test/wt-libs/on-chain-data/hotels/wt-index.spec.js
